修改监听

+++ Common/Timer/Component/TimerComponent.h
@@ -49,6 +49,6 @@ namespace Tendo
 		long long mNextUpdateTime;
         std::queue<long long> mRemoveTimers;
 		std::vector<TimeWheelLayer*> mTimerLayers;
-		std::unordered_map<long long, std::shared_ptr<TimerBase>> mTimerMap;//所有timer的列表
+		std::unordered_map<unsigned int, std::shared_ptr<TimerBase>> mTimerMap;//所有timer的列表
 	};
 }// namespace Sentry
